making it work with carSlice
