+++ client/package.json
@@ -2,7 +2,7 @@
 	"name": "client",
 	"version": "0.1.0",
 	"private": true,
-	"proxy": "http://localhost:5000",
+	"proxy": "http://localhost:5555",
 	"dependencies": {
 		"@babel/plugin-proposal-private-property-in-object": "^7.21.11",
 		"@reduxjs/toolkit": "^1.9.7",
