+++ client/src/components/CarList copy.js
@@ -6,14 +6,12 @@ import ExpandablePanel from './ExpandablePanel';
 const CarList = () => {
 	const { data, error, isLoading } = useFetchCarQuery();
 	console.log(`data is ${data}`);
-
 	let content;
-
 	if (isLoading) {
 		content = <Skeleton times={3} />;
 	} else if (error) {
-		content = <div>Error loading cars.</div>;
-	} else if (data) {
+		content = <div>Error loading albums.</div>;
+	} else {
 		content = data.map((car) => {
 			const header = <div>{car.carMakerName}</div>;
 
@@ -26,10 +24,7 @@ const CarList = () => {
 				</ExpandablePanel>
 			);
 		});
-	} else {
-		content = <div>No cars available.</div>;
 	}
-
 	return <div>{content}</div>;
 };
